refactor utils.BaseDB to sql_utils.BaseSQLDB
