+++ src/diracx/db/auth/db.py
@@ -13,7 +13,7 @@ from diracx.core.exceptions import (
     PendingAuthorizationError,
 )
 
-from ..utils import BaseDB, substract_date
+from ..sql_utils import BaseSQLDB, substract_date
 from .schema import (
     AuthorizationFlows,
     DeviceFlows,
@@ -28,7 +28,7 @@ USER_CODE_ALPHABET = "BCDFGHJKLMNPQRSTVWXZ"
 MAX_RETRY = 5
 
 
-class AuthDB(BaseDB):
+class AuthDB(BaseSQLDB):
     # This needs to be here for the BaseDB to create the engine
     metadata = AuthDBBase.metadata
