+++ src/diracx/db/dummy/db.py
@@ -4,12 +4,12 @@ from uuid import UUID
 
 from sqlalchemy import func, insert, select
 
-from ..utils import BaseDB, apply_search_filters
+from ..sql_utils import BaseSQLDB, apply_search_filters
 from .schema import Base as DummyDBBase
 from .schema import Cars, Owners
 
 
-class DummyDB(BaseDB):
+class DummyDB(BaseSQLDB):
     """
     This DummyDB is just to illustrate some important aspect of writing
     DB classes in DiracX.
